pyproject.toml
[project]
name = "telegramshablon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "environ>=1.0",
    "python-dotenv>=1.2.1",
]
